requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
Django==5.1.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
iniconfig==2.1.0
lxml==5.3.1
lxml_html_clean==0.4.1
packaging==25.0
pluggy==1.5.0
psycopg==3.2.4
PyJWT==1.7.1
pytest==8.3.5
python-dotenv==1.0.1
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
